Reorganize containers into container folder, update state correctly in StockContainer to show +/- rows. Closes #10
