+++ client/src/containers/StockContainer.jsx
@@ -1,26 +1,29 @@
 import React, { Component } from 'react';
 import Ticker from '../components/Ticker';
 import RowHeader from '../components/RowHeader';
-import RowData from '../components/RowData';
+import RowData from './RowData';
 import Price from '../components/Price';
-import RowInput from '../components/RowInput';
+import RowInput from './RowInput';
 
 class StockContainer extends Component {
     state = {
         rows: []
     }
     addRowHandler = (sentRow) => {
+        const key = this.state.rows.length;
         let rows = [...this.state.rows,
             {
                 date: sentRow.date,
                 qty: sentRow.qty,
-                cost: sentRow.cost
+                cost: sentRow.cost,
+                i: key
             }
         ]
         this.setState({ rows });
         // console.log(sentRow);
     }
-    deleteRowHandler = () => {
+    deleteRowHandler = (i) => {
+        // console.log(i)
         let rows = [...this.state.rows];
         rows.pop();
         // console.log(rows);
@@ -39,6 +42,7 @@ class StockContainer extends Component {
             qty={row.qty} 
             cost={row.cost} 
             profit={0}
+            clicked={(i) => this.deleteRowHandler(i)}
             />
         );
         console.log(this.state.rows)
